1. fix i2c & usart & debug
2. documented i2c & encoder timings
3. added printing of system information

+++ shared/delay.h
@@ -8,9 +8,9 @@
 bool Delay_Init(void);
 bool Delay_DeInit(void);
 void Delay_WaitTime_MilliSec(u32 ms);
-void Delay_WaitTime_MicroSec(u64 us);
+void Delay_WaitTime_MicroSec(u32 us);
 u32 Delay_TimeMilliSec_Get(void);
-u64 Delay_TimeMicroSec_Get(void);
+u32 Delay_TimeMicroSec_Get(void);
 double Delay_TimeAccurate_Get(void);
 void Delay_SuspendTimer(void);
 void Delay_ResumeTimer(void);
